v1.01
* git_restapi:    ca727191c3bcdb7a8cb076daae7faa7152ff4cbf
* scheduler:      75a27b46f9a28fdc60c2a4383c343eb953fe224b
* publisher:      b46adc3b80309eda81a32d0358e1caa1096dc79d
* vis_comparator: fa7efafbe560dad01989d8f2d512fef9b2676946
* add: tlspuffin, packages scheduler/git_restapi/publisher/vis_comparator plus shared
      html/tools assets into a deployable install, with per-service config generation
      and systemd samples
* change: separating tlspuffin-specific code from the generic servers
    (scheduler/publisher done; git_restapi still pending)
  * move: reserve_port from scheduler to tlspuffin (no longer puffin-specific)
  * move: client html/js assets (board launchers, publisher summary, tools/js, tlspuffin
      job scripts) from scheduler/publisher into tlspuffin/data
* move: navigation code into tlspuffin/data/html
* change: tlspuffin-specific code and navigation support package selection

+++ scheduler/src/utils/file.hxx
@@ -14,11 +14,12 @@ enum class FileReadState {
 };
 
 struct FileExtractedText {
-  ssize_t requestReadSize = 0;
-  ssize_t requestReadOffset = 0;
-  ssize_t startOffset = 0;
-  size_t filesize = 0;
-  size_t fileStartOffset = 0;
+  int64_t requestReadSize = 0;
+  int64_t requestReadOffset = 0;
+  int64_t startOffset = 0;
+  int64_t filesize = 0;
+  int64_t fileStartOffset = 0;
+
   std::string buffer;
   bool supportSeek = true;
   bool partialFile = false;
